% Auto-generated: do not edit by hand
\name{''DataFlow}

\alias{''DataFlow}

\title{DataFlow component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DataFlow(id=NULL, edges=NULL, graphType=NULL, meta=NULL,
nodeTypes=NULL, nodes=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{edges}{Unnamed list. Array of edges connecting the nodes}

\item{graphType}{Character. Type of the graph structure [singleOutput, multiPlot]}

\item{meta}{Named list. Metadata dictionary of the available databases}

\item{nodeTypes}{List of characters. Metadata dictionary of the available databases}

\item{nodes}{Unnamed list. Array of nodes}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

