% Auto-generated: do not edit by hand
\name{''DataFlow}

\alias{''DataFlow}

\title{DataFlow component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DataFlow(id=NULL, edges=NULL, meta=NULL, nodes=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{edges}{Unnamed list. The value displayed in the input.}

\item{meta}{Named list. The value displayed in the input.}

\item{nodes}{Unnamed list. A label that will be printed when this component is rendered.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

